# msa-product-receiver/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Product API
  description: Product management API for msa-product-receiver
  version: v1.0.0
  contact:
    name: Product Team
    email: product@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
tags:
  - name: Products
    description: Product management operations
paths:
  /api/v1/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve a list of all products
      operationId: getAllProducts
      responses:
        '200':
          description: List of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListProductDto'
    post:
      tags:
        - Products
      summary: Create a new product
      description: Create a new product with the provided data
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProductDto'
        '400':
          description: Invalid input
        '409':
          description: Product with same SKU already exists
  /api/v1/products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a product by its unique identifier
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProductDto'
        '404':
          description: Product not found
    put:
      tags:
        - Products
      summary: Update product
      description: Update an existing product
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseProductDto'
        '400':
          description: Invalid input
        '404':
          description: Product not found
        '409':
          description: Product with same SKU already exists
    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product by its ID
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Product ID
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
components:
  schemas:
    ProductDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier
        sku:
          type: string
          description: Stock keeping unit (article number)
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          format: decimal
          description: Price in currency
        quantity:
          type: integer
          description: Quantity in stock
        weight:
          type: number
          format: double
          description: Weight in kilograms
        isAvailable:
          type: boolean
          description: Availability flag
        category:
          $ref: '#/components/schemas/ProductCategory'
        tags:
          type: array
          items:
            type: string
          description: List of tags
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    ProductCategory:
      type: string
      enum:
        - ELECTRONICS
        - CLOTHING
        - FOOD
        - HOME_GOODS
      description: Product category enumeration
    CreateProductRequest:
      type: object
      required:
        - sku
        - name
        - description
        - price
        - quantity
        - isAvailable
        - category
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 50
          description: Stock keeping unit (article number)
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          description: Price in currency
        quantity:
          type: integer
          minimum: 0
          description: Quantity in stock
        weight:
          type: number
          format: double
          description: Weight in kilograms
        isAvailable:
          type: boolean
          description: Availability flag
        category:
          type: string
          description: Product category
        tags:
          type: array
          items:
            type: string
          description: List of tags
    UpdateProductRequest:
      type: object
      properties:
        sku:
          type: string
          minLength: 1
          maxLength: 50
          description: Stock keeping unit (article number)
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          format: decimal
          minimum: 0
          exclusiveMinimum: true
          description: Price in currency
        quantity:
          type: integer
          minimum: 0
          description: Quantity in stock
        weight:
          type: number
          format: double
          description: Weight in kilograms
        isAvailable:
          type: boolean
          description: Availability flag
        category:
          type: string
          description: Product category
        tags:
          type: array
          items:
            type: string
          description: List of tags
    ApiResponseProductDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Response message
        data:
          $ref: '#/components/schemas/ProductDto'
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
    ApiResponseListProductDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Response message
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductDto'
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the response
    ApiError:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error description
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
        path:
          type: string
          description: Path where the error occurred
security: []
externalDocs:
  description: Product API Documentation
  url: https://example.com/docs